PrefUI // Massive reconstruction.
- Also synchronize localizable.strings across distros.
