+++ Source/Modules/UIModules/PrefUI/VwrPrefPaneDictionary.swift
@@ -27,37 +27,18 @@ struct VwrPrefPaneDictionary: View {
   @State private var selEnableCNS11643: Bool = UserDefaults.standard.bool(forKey: UserDef.kCNS11643Enabled.rawValue)
   @State private var selEnableSymbolInputSupport: Bool = UserDefaults.standard.bool(
     forKey: UserDef.kSymbolInputEnabled.rawValue)
-  @State private var selAllowBoostingSingleKanjiAsUserPhrase: Bool = UserDefaults.standard.bool(
-    forKey: UserDef.kAllowBoostingSingleKanjiAsUserPhrase.rawValue)
   @State private var selFetchSuggestionsFromUserOverrideModel: Bool = UserDefaults.standard.bool(
     forKey: UserDef.kFetchSuggestionsFromUserOverrideModel.rawValue)
-  @State private var selUseFixecCandidateOrderOnSelection: Bool = UserDefaults.standard.bool(
-    forKey: UserDef.kUseFixecCandidateOrderOnSelection.rawValue)
-  @State private var selConsolidateContextOnCandidateSelection: Bool = UserDefaults.standard.bool(
-    forKey: UserDef.kConsolidateContextOnCandidateSelection.rawValue)
-  @State private var selHardenVerticalPunctuations: Bool = UserDefaults.standard.bool(
-    forKey: UserDef.kHardenVerticalPunctuations.rawValue)
+  @State private var selPhraseReplacementEnabled: Bool = UserDefaults.standard.bool(
+    forKey: UserDef.kPhraseReplacementEnabled.rawValue
+  )
 
   private static let dlgOpenPath = NSOpenPanel()
   private static let dlgOpenFile = NSOpenPanel()
 
-  private let contentMaxHeight: Double = 490
-  private let contentWidth: Double = {
-    switch PrefMgr.shared.appleLanguages[0] {
-    case "ja":
-      return 520
-    default:
-      if PrefMgr.shared.appleLanguages[0].contains("zh-Han") {
-        return 480
-      } else {
-        return 580
-      }
-    }
-  }()
-
   var body: some View {
     ScrollView {
-      SSPreferences.Container(contentWidth: contentWidth) {
+      SSPreferences.Container(contentWidth: CtlPrefUI.contentWidth) {
         // MARK: - User Data Folder Path Management
 
         SSPreferences.Section(title: "", bottomDivider: true) {
@@ -126,6 +107,12 @@ struct VwrPrefPaneDictionary: View {
                 }
               }
             ).controlSize(.small)
+            Text(
+              LocalizedStringKey(
+                "Due to security concerns, we don't consider implementing anything related to shell script execution here. An input method doing this without implementing App Sandbox will definitely have system-wide vulnerabilities, considering that its related UserDefaults are easily tamperable to execute malicious shell scripts. Official releases of vChewing is Sandboxed, and a Sandboxed app is not allowed to execute an external shell command, either."
+              )
+            )
+            .preferenceDescription()
           }
           Divider()
           Group {
@@ -141,7 +128,7 @@ struct VwrPrefPaneDictionary: View {
                 "This will use the plist files deployed by the “make install” command from libvChewing-Data if possible."
               )
             )
-            .preferenceDescription().fixedSize(horizontal: false, vertical: true)
+            .preferenceDescription()
             Toggle(
               LocalizedStringKey("Only load factory language models if needed"),
               isOn: $selOnlyLoadFactoryLangModelsIfNeeded.onChange {
@@ -162,56 +149,28 @@ struct VwrPrefPaneDictionary: View {
                 LMMgr.setSymbolEnabled(PrefMgr.shared.symbolInputEnabled)
               }
             )
-          }
-          // Divider()
-          Group {
-            Toggle(
-              LocalizedStringKey("Allow boosting / excluding a candidate of single kanji"),
-              isOn: $selAllowBoostingSingleKanjiAsUserPhrase.onChange {
-                PrefMgr.shared.allowBoostingSingleKanjiAsUserPhrase = selAllowBoostingSingleKanjiAsUserPhrase
-              }
-            )
             Toggle(
               LocalizedStringKey("Applying typing suggestions from half-life user override model"),
               isOn: $selFetchSuggestionsFromUserOverrideModel.onChange {
                 PrefMgr.shared.fetchSuggestionsFromUserOverrideModel = selFetchSuggestionsFromUserOverrideModel
               }
             )
-            Toggle(
-              LocalizedStringKey("Always use fixed listing order in candidate window"),
-              isOn: $selUseFixecCandidateOrderOnSelection.onChange {
-                PrefMgr.shared.useFixecCandidateOrderOnSelection = selUseFixecCandidateOrderOnSelection
-              }
-            )
-            Toggle(
-              LocalizedStringKey("Consolidate the context on confirming candidate selection"),
-              isOn: $selConsolidateContextOnCandidateSelection.onChange {
-                PrefMgr.shared.consolidateContextOnCandidateSelection = selConsolidateContextOnCandidateSelection
-              }
-            )
             Text(
-              LocalizedStringKey(
-                "For example: When typing “章太炎” and you want to override the “太” with “泰”, and the raw operation index range [1,2) which bounds are cutting the current node “章太炎” in range [0,3). If having lack of the pre-consolidation process, this word will become something like “張泰言” after the candidate selection. Only if we enable this consolidation, this word will become “章泰炎” which is the expected result that the context is kept as-is."
-              )
+              "The user override model only possesses memories temporarily. It won't memorize those unigrams consisting of only one Chinese character, except “你/他/妳/她/祢/衪/它/牠/再/在”. Each memory record gradually becomes ineffective within approximately less than 6 days. You can erase all memory records through the input method menu.".localized
             )
-            .preferenceDescription().fixedSize(horizontal: false, vertical: true)
+            .preferenceDescription()
             Toggle(
-              LocalizedStringKey("Harden vertical punctuations during vertical typing (not recommended)"),
-              isOn: $selHardenVerticalPunctuations.onChange {
-                PrefMgr.shared.hardenVerticalPunctuations = selHardenVerticalPunctuations
+              LocalizedStringKey("Enable phrase replacement table"),
+              isOn: $selPhraseReplacementEnabled.onChange {
+                PrefMgr.shared.phraseReplacementEnabled = selPhraseReplacementEnabled
               }
             )
-            Text(
-              LocalizedStringKey(
-                "⚠︎ This feature is useful ONLY WHEN the font you are using doesn't support dynamic vertical punctuations. However, typed vertical punctuations will always shown as vertical punctuations EVEN IF your editor has changed the typing direction to horizontal."
-              )
-            )
-            .preferenceDescription().fixedSize(horizontal: false, vertical: true)
+            Text("This will batch-replace specified candidates.".localized).preferenceDescription()
           }
         }
       }
     }
-    .frame(maxHeight: contentMaxHeight).fixedSize(horizontal: false, vertical: true)
+    .frame(maxHeight: CtlPrefUI.contentMaxHeight).fixedSize(horizontal: false, vertical: true)
   }
 }
